Improve job execution concurrency
